requirements.txt
langchain
langchain_community
langchain_text_splitters
llama-cpp-python
sentence-transformers
chromadb
pypdf
requests
tqdm
langdetect
fastapi 
uvicorn
python-multipart
python-dotenv